+++ source/pod.go
@@ -78,6 +78,7 @@ func NewPodSource(
 	informers.MustAddIndexers(podInformer.Informer(), informers.IndexerWithOptions[*v1.Pod](
 		informers.IndexSelectorWithAnnotationFilter(annotationFilter),
 		informers.IndexSelectorWithLabelSelector(labelSelector),
+		informers.IndexSelectorWithConditions(annotations.IsControllerMatch[*v1.Pod]),
 	))
 	informers.MustSetTransform(podInformer.Informer(), informers.TransformerWithOptions[*v1.Pod](
 		informers.TransformRemoveManagedFields(),
@@ -119,7 +120,7 @@ func (ps *podSource) AddEventHandler(_ context.Context, handler func()) {
 func (ps *podSource) Endpoints(_ context.Context) ([]*endpoint.Endpoint, error) {
 	indexKeys := ps.podInformer.Informer().GetIndexer().ListIndexFuncValues(informers.IndexWithSelectors)
 
-	endpoints := make([]*endpoint.Endpoint, 0)
+	endpoints := make([]*endpoint.Endpoint, 0, len(indexKeys))
 	for _, key := range indexKeys {
 		pod, err := informers.GetByKey[*v1.Pod](ps.podInformer.Informer().GetIndexer(), key)
 		if err != nil {
@@ -141,7 +142,7 @@ func (ps *podSource) Endpoints(_ context.Context) ([]*endpoint.Endpoint, error) 
 		endpoints = append(endpoints, podEndpoints...)
 	}
 
-	return MergeEndpoints(endpoints), nil
+	return endpoint.MergeEndpoints(endpoints), nil
 }
 
 func (ps *podSource) endpointsFromPodAnnotations(pod *v1.Pod) []*endpoint.Endpoint {