+++ source/pod_fqdn_test.go
@@ -172,7 +172,7 @@ func TestPodSourceFqdnTemplatingExamples(t *testing.T) {
 						Name:      "pod-1",
 						Namespace: "default",
 						Annotations: map[string]string{
-							"external-dns.alpha.kubernetes.io/target": "203.2.45.22",
+							"external-dns.kubernetes.io/target": "203.2.45.22",
 						},
 					},
 					Status: v1.PodStatus{
@@ -197,7 +197,7 @@ func TestPodSourceFqdnTemplatingExamples(t *testing.T) {
 						Name:      "pod-1",
 						Namespace: "default",
 						Annotations: map[string]string{
-							"external-dns.alpha.kubernetes.io/hostname": "ip-10-1-176-1.internal.domain.com",
+							"external-dns.kubernetes.io/hostname": "ip-10-1-176-1.internal.domain.com",
 						},
 					},
 					Status: v1.PodStatus{